# docs/swagger.yaml
basePath: /api
definitions:
...
  cmd.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  cmd.Payment:
    properties:
      amount:
        format: float64
        type: number
      createdAt:
        type: string
      currency:
        type: string
      id:
        type: string
      method:
        $ref: '#/definitions/go-ddd-clean_internal_domain_payment.Method'
      sessionID:
        type: string
      status:
        $ref: '#/definitions/go-ddd-clean_internal_domain_payment.Status'
      transactionRef:
        type: string
    type: object
...
  cmd.Photo:
    properties:
      composition:
        additionalProperties: {}
        type: object
      createdAt:
        type: string
      filterID:
        type: string
      frameID:
        type: string
      id:
        type: string
      renderedURL:
        type: string
      sessionID:
        type: string
      storageURL:
        type: string
    type: object
...
  cmd.Session:
    properties:
      boothID:
        type: string
      boothSnapshot:
        additionalProperties: {}
        type: object
      finishedAt:
        type: string
      id:
        type: string
      paymentID:
        type: string
      phoneTemp:
        type: string
      startedAt:
        type: string
      status:
        $ref: '#/definitions/go-ddd-clean_internal_domain_session.Status'
      totalPrice:
        format: float64
        type: number
      userID:
        type: string
      voucherID:
        type: string
    type: object
  cmd.SessionCreateRequest:
    properties:
      booth_id:
        type: string
      booth_snapshot:
        additionalProperties: {}
        type: object
      payment_id:
        type: string
      phone_temp:
        type: string
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: string
      voucher_id:
        type: string
    type: object
  cmd.SessionUpdateRequest:
    properties:
      booth_snapshot:
        additionalProperties: {}
        type: object
      finished_at:
        type: integer
      payment_id:
        type: string
      phone_temp:
        type: string
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: string
      voucher_id:
        type: string
    type: object
  cmd.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      password:
        type: string
      points:
        type: integer
      role:
        $ref: '#/definitions/go-ddd-clean_internal_domain_user.Role'
      tel:
        type: string
      updatedAt:
        type: string
    type: object
  cmd.UserAdjustPointsRequest:
    properties:
      delta:
        type: integer
    type: object
  cmd.UserCreateRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
      tel:
        type: string
    type: object
  cmd.UserUpdateRequest:
    properties:
      email:
        type: string
      password:
        type: string
      points:
        type: integer
      role:
        type: string
      tel:
        type: string
    type: object
...
  go-ddd-clean_internal_domain_payment.Method:
    enum:
    - cash
    - qr
    - stripe
    - points
    type: string
    x-enum-varnames:
    - MethodCash
    - MethodQR
    - MethodStripe
    - MethodPoints
  go-ddd-clean_internal_domain_payment.Status:
    enum:
    - pending
    - success
    - failed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusSuccess
    - StatusFailed
  go-ddd-clean_internal_domain_session.Status:
    enum:
    - started
    - success
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - StatusStarted
    - StatusSuccess
    - StatusFailed
    - StatusCancelled
  go-ddd-clean_internal_domain_user.Role:
    enum:
    - customer
    - staff
    - admin
    type: string
    x-enum-varnames:
    - RoleCustomer
    - RoleStaff
    - RoleAdmin
...
info:
  contact: {}
  description: API documentation for the go-ddd-clean service.
  title: Photobooth Platforms API
  version: "1.0"
paths:
...
  /api/payments/session/{sessionID}:
    get:
      parameters:
      - description: รหัสเซสชัน
        in: path
        name: sessionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Payment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดูข้อมูลการชำระเงินของเซสชัน
      tags:
      - Payments
  /api/sessions:
    get:
      parameters:
      - description: กรองตามสถานะ
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.Session'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: ดึงรายการเซสชัน
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลเซสชัน
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.SessionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: สร้างเซสชัน
      tags:
      - Sessions
  /api/sessions/{id}:
    delete:
      parameters:
      - description: รหัสเซสชัน
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: ลบเซสชัน
      tags:
      - Sessions
    get:
      parameters:
      - description: รหัสเซสชัน
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Session'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: ดูข้อมูลเซสชัน
      tags:
      - Sessions
    put:
      consumes:
      - application/json
      parameters:
      - description: รหัสเซสชัน
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลที่ต้องการแก้ไข
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.SessionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Session'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: ปรับปรุงข้อมูลเซสชัน
      tags:
      - Sessions
  /api/sessions/{id}/payment:
    get:
      parameters:
      - description: รหัสเซสชัน
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Payment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: ดูข้อมูลการชำระเงินของเซสชัน
      tags:
      - Sessions
  /api/sessions/{id}/photos:
    get:
      parameters:
      - description: รหัสเซสชัน
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.Photo'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: ดึงรายการรูปของเซสชัน
      tags:
      - Sessions
  /api/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดึงรายการผู้ใช้
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลผู้ใช้
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: สร้างผู้ใช้ใหม่
      tags:
      - Users
  /api/users/{id}:
    delete:
      parameters:
      - description: รหัสผู้ใช้
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ลบผู้ใช้
      tags:
      - Users
    get:
      parameters:
      - description: รหัสผู้ใช้
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดูข้อมูลผู้ใช้
      tags:
      - Users
    put:
      consumes:
      - application/json
      parameters:
      - description: รหัสผู้ใช้
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลที่ต้องแก้ไข
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ปรับปรุงข้อมูลผู้ใช้
      tags:
      - Users
  /api/users/{id}/points:
    post:
      consumes:
      - application/json
      parameters:
      - description: รหัสผู้ใช้
        in: path
        name: id
        required: true
        type: string
      - description: ค่าที่ต้องการปรับ
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.UserAdjustPointsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ปรับแต้มผู้ใช้
      tags:
      - Users
...
schemes:
- http
securityDefinitions:
  BoothTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"